Relocate the App component to the components folder
